feat(logger): remove OutputChannel creation and add as optional
Remove OutputChannel creation and add OutputChannel as an optional parameter. It will allow the VS
Code Extension to write logs to the files only and not to Output channel.

BREAKING CHANGE: OutputChannel will not be created inside the logger as today. Alternatively, it
will be added as an optional parameter. The VS Code Extension will be able to create the
OutputChannel and send it to the logger as a parameter.

fix #55