feat(logger): a memory leak when creating 9 or more ChildLogger's
Fix a memory leak caused by creating winston logger with the same transport for each ChildLogger.
Change the internal implementation to use the same Winston logger for all ChildLogger's.
Add namespace field to all the log messages which will show the correct label (the full logger name).

BREAKING CHANGE inside the log messages:
A new namespace field is added to all the log messages.
It shows the correct label (the full logger name) and can not be overwritten by the log methods.

fix #76